// games/pirate-nation-showcase/src/avatar/avatarCatalog.generated.ts
export const AVATAR_SLOTS = ["species","face","eyebrow","hair","facialhair","ears","eyewear","headwear","tops","bottoms","shoes","back"] as const

export type AvatarSlot = (typeof AVATAR_SLOTS)[number]